edify_backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
axios==0.4.0
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.29
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
idna==3.11
itypes==1.2.0
Jinja2==3.1.6
kombu==5.6.2
lxml==6.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
Pygments==2.20.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
redis==7.0.1
requests==2.32.5
rich==14.3.3
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
PyVimeo==1.1.2
google-api-python-client
google-auth-oauthlib
# Deployment additions (Docker / prod)
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
sentry-sdk==2.19.2
